+++ Assets/Scripts/Engine/Inventory/InventoryEntry.cs.meta
